Migrate diffsky phot generator to experimental:
